Split compare detail page helpers
